# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Social Distribution - Ctrl + Alt + Defeat API",
      "description": "API for our backend"
    },
    "paths": {
...
          "/api/follow/usersearch": {
            "get": {
              "tags": ["Follow"],
              "summary": "Search for users",
              "parameters": [
                {
                  "name": "query",
                  "in": "query",
                  "type": "string",
                  "description": "Search query",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "users": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "username": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "/api/follow/follow_request": {
            "post": {
              "tags": ["Follow"],
              "summary": "Send follow request",
              "parameters": [
                {
                  "name": "body",
                  "in": "body",
                  "required": true,
                  "schema": {
                    "type": "object",
                    "properties": {
                      "author_send": {
                        "type": "string"
                      },
                      "author_receive": {
                        "type": "string"
                      }
                    }
                  }
                }
              ],
              "responses": {
                "200": {
                  "description": "Follow request sent",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "/api/follow/show_requests": {
            "get": {
              "tags": ["Follow"],
              "summary": "Get follow requests",
              "parameters": [
                {
                  "name": "authorId",
                  "in": "query",
                  "type": "string",
                  "description": "Author ID",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "followRequests": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "username": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "/api/follow/accept_request": {
            "post": {
              "tags": ["Follow"],
              "summary": "Accept follow request",
              "parameters": [
                {
                  "name": "body",
                  "in": "body",
                  "required": true,
                  "schema": {
                    "type": "object",
                    "properties": {
                      "author_followee": {
                        "type": "string"
                      },
                      "author_following": {
                        "type": "string"
                      }
                    }
                  }
                }
              ],
              "responses": {
                "200": {
                  "description": "Follow request accepted",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "/api/follow/reject_request": {
            "post": {
              "tags": ["Follow"],
              "summary": "Reject follow request",
              "parameters": [
                {
                  "name": "body",
                  "in": "body",
                  "required": true,
                  "schema": {
                    "type": "object",
                    "properties": {
                      "author_followee": {
                        "type": "string"
                      },
                      "author_following": {
                        "type": "string"
                      }
                    }
                  }
                }
              ],
              "responses": {
                "200": {
                  "description": "Follow request rejected",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "/api/follow/unfollow": {
            "post": {
              "tags": ["Follow"],
              "summary": "Unfollow a user",
              "parameters": [
                {
                  "name": "body",
                  "in": "body",
                  "required": true,
                  "schema": {
                    "type": "object",
                    "properties": {
                      "author_unfollow": {
                        "type": "string"
                      },
                      "author_unfollower": {
                        "type": "string"
                      }
                    }
                  }
                }
              ],
              "responses": {
                "200": {
                  "description": "Unfollowed successfully",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "/api/authors/{author_id}/followers": {
            "get": {
              "tags": ["Follow"],
              "summary": "Get followers (REMOTE)",
              "parameters": [
                {
                  "name": "author_id",
                  "in": "path",
                  "type": "string",
                  "description": "Author ID",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string"
                      },
                      "items": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "type": {
                              "type": "string"
                            },
                            "id": {
                              "type": "string"
                            },
                            "url": {
                              "type": "string"
                            },
                            "host": {
                              "type": "string"
                            },
                            "displayName": {
                              "type": "string"
                            },
                            "github": {
                              "type": "string"
                            },
                            "profileImage": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "/api/authors/{author_id}/followers/{foreign_author_id}": {
            "put": {
              "tags": ["Follow"],
              "summary": "Add follower",
              "parameters": [
                {
                  "name": "author_id",
                  "in": "path",
                  "type": "string",
                  "description": "Author ID",
                  "required": true
                },
                {
                  "name": "foreign_author_id",
                  "in": "path",
                  "type": "string",
                  "description": "Foreign Author ID",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Follower added",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "get": {
              "tags": ["Follow"],
              "summary": "Check if user is a follower (REMOTE)",
              "parameters": [
                {
                  "name": "author_id",
                  "in": "path",
                  "type": "string",
                  "description": "Author ID",
                  "required": true
                },
                {
                  "name": "foreign_author_id",
                  "in": "path",
                  "type": "string",
                  "description": "Foreign Author ID",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Successful response",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "is_follower": {
                        "type": "boolean"
                      }
                    }
                  }
                }
              }
            },
            "delete": {
              "tags": ["Follow"],
              "summary": "Remove follower",
              "parameters": [
                {
                  "name": "author_id",
                  "in": "path",
                  "type": "string",
                  "description": "Author ID",
                  "required": true
                },
                {
                  "name": "foreign_author_id",
                  "in": "path",
                  "type": "string",
                  "description": "Foreign Author ID",
                  "required": true
                }
              ],
              "responses": {
                "200": {
                  "description": "Follower removed",
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
        }
      }
    }
  }